fix: изменение типа данных

+++ src/OsuNet/Models/UserBest.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 using OsuNet.Enums;
 
 namespace OsuNet.Models {
@@ -100,7 +101,7 @@ namespace OsuNet.Models {
         /// The number of PP that the player has hit.
         /// </summary>
         [JsonProperty("pp")]
-        public double PP { get; set; }
+        public float PP { get; set; }
 
         /// <summary>
         /// True if you can watch the replay, otherwise false.
